+++ user/lib/string.h
@@ -10,7 +10,7 @@ void *memcpy(void *dest, const void *src, size_t n);
 void *memmove(void *dest, const void *src, size_t n);
 void *memset(void *dest, int ch, size_t n);
 int strncmp(const char *s1, const char *s2, size_t n);
-char *strncpy(char *s, const char *t, size_t n);
+char *strncpy(char *s, const char *t, int n);
 size_t strlen(const char *s);
 }
